Refactor: 파일 이름을 좀 더 명확하게 수정 #103
